Fix shortcut icon URL
Before, the default shortcut icon was assumed relative to the SassDoc
configuration file (or CWD), which resulted in a missing file. Though,
since the assets directory is always copied, the shortcut icon was still
in the expected place.

This commit fixes the behavior; when no shortcut icon is specified,
don't try do copy the default one, just set the URL.

Also, custom icons are copied in `images` directory instead of `img`
directory for consistency.